+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 ==========
 
+Unreleased
+**********
+Added: CalculatorIcon and options ability for header to display a Sensitivity Calculator button.
+
 2.8.40
 ******